# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ODIN Swagger UI",
        "title": "Odin: Remote Administration for 1S.Enterprise Application Servers",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "khorevaa@yandex.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/api/v1",
    "paths": {
...
        "/app/{app}/clusters/{cluster}/processes": {
            "get": {
                "description": "получение списка процессов на кластере",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processes"
                ],
                "summary": "получение списка процессов на кластере",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app name",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster uuid",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster user",
                        "name": "cluster-usr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster password",
                        "name": "cluster-pwd",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "force update ignore cache",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/serialize.ProcessInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
        "/app/{app}/clusters/{cluster}/processes/{process}": {
            "get": {
                "description": "получение информации опроцессе на кластере",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processes"
                ],
                "summary": "получение информации опроцессе на кластере",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app name",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster uuid",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "uuid process",
                        "name": "process",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster user",
                        "name": "cluster-usr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster password",
                        "name": "cluster-pwd",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/serialize.ProcessInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
...
        "/app/{app}/processes": {
            "get": {
                "description": "получение списка процессов на сервере 1С Предприятие",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processes"
                ],
                "summary": "получение списка процессов на сервере 1С Предприятие",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app name",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster uuid",
                        "name": "cluster-id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster user",
                        "name": "cluster-usr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster password",
                        "name": "cluster-pwd",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "force update ignore cache",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/serialize.ProcessInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
        "/app/{app}/processes/{process}": {
            "get": {
                "description": "получение информации опроцессе на сервере 1С Предприятие",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processes"
                ],
                "summary": "получение информации опроцессе на сервере 1С Предприятие",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app name",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "uuid process",
                        "name": "process",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster uuid",
                        "name": "cluster-id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cluster user",
                        "name": "cluster-usr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster password",
                        "name": "cluster-pwd",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/serialize.ProcessInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "serialize.LicenseInfo": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string",
                    "example": "1CV8"
                },
                "full_name": {
                    "type": "string",
                    "example": "FullName"
                },
                "full_presentation": {
                    "type": "string",
                    "example": "Сервер, 2300, app, 1541, ORG8A Сетевой 300"
                },
                "host": {
                    "type": "string",
                    "example": "host"
                },
                "issued_by_server": {
                    "type": "boolean",
                    "example": true
                },
                "license_type": {
                    "description": "license-type       : HASP",
                    "type": "integer",
                    "example": 0
                },
                "max_users_all": {
                    "type": "integer",
                    "example": 300
                },
                "max_users_cur": {
                    "type": "integer",
                    "example": 300
                },
                "net": {
                    "type": "boolean",
                    "example": true
                },
                "process_id": {
                    "type": "string",
                    "example": "94232f94-be78-4acd-a11e-09911bd4f4ed"
                },
                "rmngr_address": {
                    "type": "string",
                    "example": "app"
                },
                "rmngr_pid": {
                    "type": "string",
                    "example": "2300"
                },
                "rmngr_port": {
                    "type": "integer",
                    "example": 1541
                },
                "series": {
                    "type": "string",
                    "example": "ORG8A"
                },
                "session_id": {
                    "type": "string",
                    "example": "e45c1c2b-b3ac-4fea-9f0c-0583ad65d117"
                },
                "short_presentation": {
                    "type": "string",
                    "example": "Сервер, ORG8A Сет 300"
                },
                "user_name": {
                    "type": "string",
                    "example": "User"
                }
            }
        },
...
        "serialize.ProcessInfo": {
            "type": "object",
            "properties": {
                "available_perfomance": {
                    "type": "integer",
                    "example": 100
                },
                "avg_back_call_time": {
                    "type": "number",
                    "example": 0
                },
                "avg_call_time": {
                    "type": "number",
                    "example": 0.483
                },
                "avg_db_call_time": {
                    "type": "number",
                    "example": 0.124
                },
                "avg_lock_call_time": {
                    "type": "number",
                    "example": 0
                },
                "avg_server_call_time": {
                    "type": "number",
                    "example": -0.265
                },
                "avg_threads": {
                    "type": "number",
                    "example": 0.281
                },
                "capacity": {
                    "type": "integer",
                    "example": 1000
                },
                "cluster_id": {
                    "type": "string",
                    "example": "0e588a25-8354-4344-b935-53442312aa30"
                },
                "connections": {
                    "type": "integer",
                    "example": 7
                },
                "enable": {
                    "type": "boolean",
                    "example": true
                },
                "host": {
                    "type": "string",
                    "example": "srv"
                },
                "licenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/serialize.LicenseInfo"
                    }
                },
                "memory_excess_time": {
                    "type": "integer",
                    "example": 0
                },
                "memory_size": {
                    "type": "integer",
                    "example": 1518604
                },
                "pid": {
                    "type": "string",
                    "example": "3366"
                },
                "port": {
                    "type": "integer",
                    "example": 1564
                },
                "reverse": {
                    "type": "boolean",
                    "example": true
                },
                "running": {
                    "type": "boolean",
                    "example": true
                },
                "selection_size": {
                    "type": "integer",
                    "example": 61341
                },
                "started_at": {
                    "type": "string",
                    "example": "2018-03-29T11:16:02"
                },
                "use": {
                    "type": "boolean",
                    "example": true
                },
                "uuid": {
                    "type": "string",
                    "example": "0e588a25-8354-4344-b935-53442312aa30"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "metrics": " Grants read access and control metrics operations",
                "read": " Grants read access",
                "write": " Grants write access"
            }
        }
    }
}